Mouse and Pointer events now contain a more props
A lot more properties have been added to the MouseEvent and PointerEvent
objects to bring them in line with most of the JS equivalents. This
includes things like special keys being down at the time of the event,
and which buttons were down at the time of the event, as well as the type
of pointer event received.