Adds JS events during engine load
While the game engine is loading the player is presented with a blank screen.
This feature allows developers to receive events from Indigo while it's
loading, allowing them to display a loading screen or splash screen.

As we also emit errors, it will also allow the developer to detect errors
and then gracefully display an error screen.

Closes #929